+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 -   Update `<DefinitionListItem />` and `<Breadcrumbs />` core components to truncate long strings
+-   Fix `<InputNumber >` core component to trigger `onChange` callback on + / - button click
 
 ## [1.0.50] - 2024-10-23